% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FishnetFunction.R
\name{FishnetFunction}
\alias{FishnetFunction}
\title{Fishnet Tool}
\usage{
FishnetFunction(my_poly, extent_only, diff_factor, length_factor)
}
\arguments{
\item{my_poly}{A sp objetct.}

\item{extent_only}{(optional) TRUE (default) or FALSE. If the user input file is a polygon,
the extent_only can be set to FALSE and the ouptut fishnet polygon will
be cropped to the input file.}

\item{diff_factor}{Integer. Faktor after which the extent of the input file will be divided
into x and y direction.}

\item{length_factor}{Integer. Can be used instead of the diff_factor. Distance in meters of
one tile n x and y direction.}
}
\value{
A fishnet polygon in UTM projection.
}
\description{
This function creates a fishnet polygon from an input shapefile. The output fishnet will
be reprojected to UTM.
}
\examples{
library(sp)
library(raster)
library(rgdal)
library(rgeos)

# Load sample shapefile including the borders of vietnam
vnm_shp <- raster::shapefile(system.file(package = "SpatialDataToolbox", "extdata",
                                         "vietnam_borders.shp"))

# Apply diff_factor
x1 <- FishnetFunction(my_poly = vnm_shp, extent_only = FALSE, diff_factor = 10)

# Apply length_factor
x2 <- FishnetFunction(my_poly = vnm_shp, extent_only = FALSE, length_factor = 100000)

}
