% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHPDissolve.R
\name{SHPDissolve}
\alias{SHPDissolve}
\title{Shapefile Dissolve Tool}
\usage{
SHPDissolve(my_poly, my_col)
}
\arguments{
\item{my_poly}{Either a SpatialPolygonsDataFrame, SpatialPointsDataFrame or SpatialLinesDataFrame.}

\item{my_col}{The attribute column after which the shapefile should be dissolved.}
}
\value{
The dissolved SpatialPolygonsDataFrame.
}
\description{
This function automatically dissolves a shapefile based on its feature attributes.
}
\examples{
library(sp)
library(raster)
library(rgdal)
library(rgeos)

# Create test polygons from extents
my_extent1 <- extent(5,7,8,10)
my_extent2 <- extent(7,9,8,10)
my_extent3 <- extent(9,11,8,10)

# Convert extents into "SpatialPolygons"
my_poly1 <- as(my_extent1, 'SpatialPolygons')
my_poly2 <- as(my_extent2, 'SpatialPolygons')
my_poly3 <- as(my_extent3, 'SpatialPolygons')

# Create dataframes
data1 <- data.frame(Id=1)
data2 <- data.frame(Id=2)

# Add the data.drames to the spatial polyongs
spdf1 = SpatialPolygonsDataFrame(my_poly1, data1)
spdf2 = SpatialPolygonsDataFrame(my_poly2, data1)
spdf3 = SpatialPolygonsDataFrame(my_poly3, data2)

# Combine all polygons
spdf_comb <- rbind(spdf1,spdf2,spdf3)

# Dissolve based on Id value
spdf_diss <- SHPDissolve(my_poly = spdf_comb, my_col = "Id")

plot(spdf_comb)
plot(spdf_diss)

}
