% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialDataToolbox.R
\docType{package}
\name{SpatialDataToolbox}
\alias{SpatialDataToolbox}
\alias{SpatialDataToolbox-package}
\title{Toolbox for Spatial Data Manipulation}
\description{
This R package contains several functions which are designed to manipulate spatial data.
This includes dissolving shapefiles based on feature attributes, converting the projection
of raster and shapefiles to UTM and subsetting data into smaller tiles.
}
\details{
Functions included in this package are:
\itemize{
  \item \code{\link{UTMConversion}}
  \item \code{\link{SHPDissolve}}
  \item \code{\link{FishnetFunction}}
  \item \code{\link{TilesMaker}}
}
}
\author{
Marius Philipp \email{marius.philipp@stud-mail.uni-wuerzburg.de}
}
