% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TilesMaker.R
\name{TilesMaker}
\alias{TilesMaker}
\title{Tiles Maker Tool}
\usage{
TilesMaker(user_file, user_aoi, diff_factor, length_factor, out_folder)
}
\arguments{
\item{user_file}{A raster or sp objetct.}

\item{user_aoi}{(optional) A sp object. Will be used to crop input data and create tiles.}

\item{diff_factor}{Integer. Faktor after which the extent of the input file will be divided
into x and y direction.}

\item{length_factor}{Integer. Can be used instead of the diff_factor. Distance in meters of
one tile in x and y direction.}

\item{out_folder}{Character string. Path to which the output tiles and fishnet polygon
should be written.}
}
\value{
A fishnet polygon in UTM projection.
}
\description{
This function divides the input file in tiles based on user preferences.
The output data will be reprojected to UTM.
}
\examples{
library(sp)
library(raster)
library(rgdal)
library(geosphere)

# Load sample raster file
my_raster <- raster::brick(system.file(package = "SpatialDataToolbox", "extdata", "landsat_sample.tif"))

# Apply diff_factor
x1 <- TilesMaker(user_file = my_raster, diff_factor = 5, out_folder = "./")

# Delete created folder with tiles and fishnet polygon
unlink("./Fishnet", recursive = TRUE)

# Apply length_factor
x2 <- TilesMaker(user_file = my_raster, length_factor = 10000, out_folder = "./")


}
