% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTMConversion.R
\name{UTMConversion}
\alias{UTMConversion}
\title{UTM Conversion Tool}
\usage{
UTMConversion(user_file)
}
\arguments{
\item{user_file}{Can either be a sp object or a raster file.}
}
\value{
The user_file in UTM projection.
}
\description{
This function automatically converts a raster or shapefile from a longlat
projection into the corresponding UTM projection.
}
\examples{
library(sp)
library(raster)
library(rgdal)

# Create a test polygon within Nigeria
my_extent <- extent(5,7,8,10)
my_poly <- as(my_extent, 'SpatialPolygons')
my_proj <- "+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0"
proj4string(my_poly) <- CRS(my_proj)

# Apply the UTM conversion function
my_poly_utm <- UTMConversion(my_poly)
crs(my_poly_utm)

}
