% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BookletMaker.R
\name{BookletMaker}
\alias{BookletMaker}
\title{Map Booklet Maker}
\usage{
BookletMaker(user_ggplot, user_fishnet, fishnet_col, user_width,
  user_height, user_pointsize, user_buffer, out_dir)
}
\arguments{
\item{user_ggplot}{A gg object. Input plot which will be used as a basis for the booklet.}

\item{user_fishnet}{A sp object. For each tile of the fishnet polygon a submap will be created.
See \code{\link{FishnetFunction}}.}

\item{fishnet_col}{(optional) Colour of the fishnet polygon within the overview map}

\item{user_width}{(optional) Width of the output pdf. Default is 7.}

\item{user_height}{(optional) Height of the output pdf. Default is 7.}

\item{user_pointsize}{(optional) The pointsize to be used. Default is 12.}

\item{user_buffer}{(optional) Buffer around each tile in m. Default is 0.}

\item{out_dir}{A character string. The directory to which the output pdf should be written.}
}
\value{
A gg object containing the original ggplot with the fishnet polygon added on top.
}
\description{
This function uses a fishnet polygon to create a map booklet pdf based on a ggplot.
The booklet contains a overview map with the original ggplot and the fishnet polygon
added on top, as well as sub-maps for each tile of the fishnet polygon.
}
\examples{
library(sp)
library(raster)
library(ggplot2)
library(geosphere)
library(RStoolbox)
library(pdftools)

# Load sample raster file
my_raster <- raster::brick(system.file(package = "SpatialDataToolbox",
                           "extdata", "landsat_sample.tif"))

# Create fishnet polygon from extent
my_extent <- as(raster::extent(my_raster), 'SpatialPolygons')
proj4string(my_extent) <- sp::CRS(as.character(raster::crs(my_raster)))
my_fishnet <- FishnetFunction(my_poly = my_extent, extent_only = TRUE, diff_factor = 4)

# Use ggRGB to make RGB plot of raster
my_plot <- RStoolbox::ggRGB(my_raster, r = 3, g = 2, b = 1, stretch = "lin") +
  ggtitle("Landsat-5 RGB") + # Add title
  theme(plot.title = element_text(hjust = 0.5, face="bold", size=14), # Adjust title
        axis.text.y = element_text(angle = 90, # Vertical y axis labels
                                   hjust = 0.5)) + # Adjust position of axis labels
  scale_x_continuous(expand = c(0,0)) + # Remove expansion in x direction
  scale_y_continuous(expand = c(0,0)) + # Remove expansion in y direction
  xlab("") + # Edit x lab
  ylab("") # Edit y lab

# Plot the ggplot
my_plot

# Execute with all parameters defined.
# In orer to have a landscape DinA4 format, multiply the width by 1.414286.
test_complex <- BookletMaker(user_ggplot = my_plot,
                             user_fishnet = my_fishnet,
                             fishnet_col <- "red",
                             user_width  = 7 * 1.414286,
                             user_height = 7,
                             user_pointsize = NULL,
                             out_dir = "./",
                             user_buffer = 0)

}
